backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
alembic==1.13.2
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
langchain==0.3.1
langchain-openai==0.2.1
httpx>=0.25.0
jieba==0.42.1
python-docx==1.1.2
PyPDF2==3.0.1
ebooklib==0.18
beautifulsoup4==4.12.3
pdf2image==1.17.0
pytesseract==0.3.13
Pillow==11.0.0
pydantic-settings==2.5.2
aiofiles==24.1.0
